+++ tests/integration/test_web_data_source.py
@@ -18,11 +18,11 @@ from web_app.data_source import (  # isort:skip  # noqa # pylint: disable=E0401,
 
 @pytest.fixture(autouse=True)
 def clear_all_cachetools_func_caches() -> Iterator[None]:
-    """Clear all `cachetools.func` caches, everywhere"""
+    """Clear all `cachetools.func` caches, everywhere."""
     yield
     institution_info._cached_get_institutions_infos.cache_clear()  # type: ignore[attr-defined]
     tc.TableConfigParser._cached_get_configs.cache_clear()  # type: ignore[attr-defined]
-    web_app.utils.oidc_tools.CurrentUser._cached_get_info.cache_clear()  # type: ignore[attr-defined]
+    web_app.data_source.oidc_tools.CurrentUser._cached_get_info.cache_clear()  # type: ignore[attr-defined]
 
 
 class TestTableConfig:
