+++ web_app/contents/wbs_generic_admin_callbacks.py
@@ -13,10 +13,10 @@ from ..config import app
 from ..data_source import data_source as src
 from ..data_source import institution_info
 from ..data_source import table_config as tc
+from ..data_source.oidc_tools import CurrentUser
 from ..data_source.utils import DataSourceException
 from ..utils import dash_utils as du
 from ..utils import types, utils
-from ..utils.oidc_tools import CurrentUser
 
 _CHANGES_COL: Final[str] = "Changes"
