+++ web_app/data_source/__init__.py
@@ -1,3 +1,9 @@
 """Init."""
 
-from . import data_source, institution_info, table_config, utils  # noqa: F401
+from . import (  # noqa: F401
+    data_source,
+    institution_info,
+    oidc_tools,
+    table_config,
+    utils,
+)
